+++ core-bundle/src/Resources/contao/classes/Versions.php
@@ -323,6 +323,23 @@ class Versions extends \Controller
 			$data[$k] = \Widget::getEmptyValueByFieldType($GLOBALS['TL_DCA'][$this->strTable]['fields'][$k]['sql']);
 		}
 
+		// Reset unique fields if the restored value already exists (see #698)
+		foreach ($data as $k=>$v)
+		{
+			if (!isset($GLOBALS['TL_DCA'][$this->strTable]['fields'][$k]['eval']['unique']) || $GLOBALS['TL_DCA'][$this->strTable]['fields'][$k]['eval']['unique'] !== true)
+			{
+				continue;
+			}
+
+			$objResult = $this->Database->prepare("SELECT COUNT(*) AS cnt FROM " . $this->strTable . " WHERE " . Database::quoteIdentifier($k) . "=?")
+										->execute($v);
+
+			if ($objResult->cnt > 0)
+			{
+				$data[$k] = Widget::getEmptyValueByFieldType($GLOBALS['TL_DCA'][$this->strTable]['fields'][$k]['sql']);
+			}
+		}
+
 		$this->Database->prepare("UPDATE " . $this->strTable . " %s WHERE id=?")
 					   ->set($data)
 					   ->execute($this->intPid);